# komga-rust/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Komga Rust API",
    "version": "0.1.0",
    "description": "Rust implementation of Komga media server for comics/manga. REST API for managing libraries, series, books, read progress, readlists, collections, and more."
  },
  "servers": [
    {
      "url": "/",
      "description": "Default server"
    }
  ],
  "paths": {
...
    "/api/v1/collections": {
      "get": {
        "tags": ["Collections"],
        "summary": "List all collections",
        "operationId": "listCollections",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of collections",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollectionPageDto" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Collections"],
        "summary": "Create a collection",
        "operationId": "createCollection",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCollectionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollectionDto" }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{id}": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get collection by ID",
        "operationId": "getCollection",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Collection details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollectionDto" }
              }
            }
          },
          "404": { "description": "Collection not found" }
        }
      },
      "patch": {
        "tags": ["Collections"],
        "summary": "Update collection",
        "operationId": "updateCollection",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCollectionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollectionDto" }
              }
            }
          },
          "404": { "description": "Collection not found" }
        }
      },
      "delete": {
        "tags": ["Collections"],
        "summary": "Delete collection",
        "operationId": "deleteCollection",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Collection deleted" },
          "404": { "description": "Collection not found" }
        }
      }
    },
    "/api/v1/collections/{id}/series": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get series in a collection",
        "operationId": "getCollectionSeries",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "List of series",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SeriesDto" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookMetadataAggregationDto": {
        "type": "object",
        "properties": {
          "booksCount": { "type": "integer" },
          "booksInProgressCount": { "type": "integer" },
          "booksReadCount": { "type": "integer" },
          "booksUnreadCount": { "type": "integer" }
        }
      },
      "SeriesDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "url": { "type": "string" },
          "libraryId": { "type": "string", "format": "uuid" },
          "booksCount": { "type": "integer" },
          "booksInProgressCount": { "type": "integer" },
          "booksReadCount": { "type": "integer" },
          "booksUnreadCount": { "type": "integer" },
          "oneshot": { "type": "boolean" },
          "deleted": { "type": "boolean" },
          "created": { "type": "string", "format": "date-time" },
          "lastModified": { "type": "string", "format": "date-time" },
          "fileLastModified": { "type": "string", "format": "date-time", "nullable": true },
          "booksMetadata": { "$ref": "#/components/schemas/BookMetadataAggregationDto" },
          "metadata": { "$ref": "#/components/schemas/SeriesMetadataDto" }
        }
      },
...
      "SeriesMetadataDto": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "statusLock": { "type": "boolean" },
          "title": { "type": "string" },
          "titleLock": { "type": "boolean" },
          "titleSort": { "type": "string" },
          "titleSortLock": { "type": "boolean" },
          "publisher": { "type": "string" },
          "publisherLock": { "type": "boolean" },
          "readingDirection": { "type": "string" },
          "readingDirectionLock": { "type": "boolean" },
          "ageRating": { "type": "integer", "nullable": true },
          "ageRatingLock": { "type": "boolean" },
          "summary": { "type": "string" },
          "summaryLock": { "type": "boolean" },
          "language": { "type": "string" },
          "languageLock": { "type": "boolean" },
          "genres": { "type": "array", "items": { "type": "string" } },
          "genresLock": { "type": "boolean" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "tagsLock": { "type": "boolean" },
          "totalBookCount": { "type": "integer", "nullable": true },
          "totalBookCountLock": { "type": "boolean" },
          "sharingLabels": { "type": "array", "items": { "type": "string" } },
          "sharingLabelsLock": { "type": "boolean" },
          "links": { "type": "array", "items": { "$ref": "#/components/schemas/MetadataLinkDto" } },
          "linksLock": { "type": "boolean" },
          "alternateTitles": { "type": "array", "items": { "$ref": "#/components/schemas/AlternateTitleDto" } },
          "alternateTitlesLock": { "type": "boolean" },
          "created": { "type": "string", "format": "date-time" },
          "lastModified": { "type": "string", "format": "date-time" }
        }
      },
      "MetadataLinkDto": {
        "type": "object",
        "properties": {
          "label": { "type": "string" },
          "url": { "type": "string", "format": "uri" }
        }
      },
      "AlternateTitleDto": {
        "type": "object",
        "properties": {
          "label": { "type": "string" },
          "title": { "type": "string" }
        }
      },
...
      "CollectionDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "ordered": { "type": "boolean" },
          "filtered": { "type": "boolean" },
          "seriesIds": { "type": "array", "items": { "type": "string", "format": "uuid" } },
          "createdDate": { "type": "string", "format": "date-time" },
          "lastModifiedDate": { "type": "string", "format": "date-time" }
        }
      },
      "CollectionPageDto": {
        "type": "object",
        "properties": {
          "content": { "type": "array", "items": { "$ref": "#/components/schemas/CollectionDto" } },
          "totalElements": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "number": { "type": "integer" },
          "size": { "type": "integer" },
          "empty": { "type": "boolean" },
          "first": { "type": "boolean" },
          "last": { "type": "boolean" },
          "numberOfElements": { "type": "integer" }
        }
      },
      "CreateCollectionRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "ordered": { "type": "boolean" },
          "seriesIds": { "type": "array", "items": { "type": "string", "format": "uuid" } }
        }
      },
      "UpdateCollectionRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "ordered": { "type": "boolean" },
          "seriesIds": { "type": "array", "items": { "type": "string", "format": "uuid" } }
        }
      },
...
    }
  },
...
}